👌 IMPROVE: refactored components
