+++ src/App.js
@@ -1,16 +1,9 @@
 import React, { useState } from "react"
 import { BrowserRouter, Route, Switch } from "react-router-dom"
-import { createMuiTheme, ThemeProvider } from "@material-ui/core"
-import { orange, red } from "@material-ui/core/colors"
-import Loginpage from "./pages/Loginpage"
-import TodoPage from "./pages/Todopage"
-
-const theme = createMuiTheme({
-  palette: {
-    primary: orange,
-    secondary: red,
-  },
-})
+import { ThemeProvider } from "@material-ui/core"
+import { theme } from "./theme/theme"
+import Login from "./Components/Login"
+import Todos from "./pages/Todos"
 
 function App() {
   const [isAdmin, setIsAdmin] = useState(false)
@@ -20,10 +13,10 @@ function App() {
       <BrowserRouter>
         <Switch>
           <Route exact path="/">
-            <Loginpage isAdmin={isAdmin} setIsAdmin={setIsAdmin} />
+            <Login isAdmin={isAdmin} setIsAdmin={setIsAdmin} />
           </Route>
           <Route path="/todo">
-            <TodoPage isAdmin={isAdmin} />
+            <Todos isAdmin={isAdmin} />
           </Route>
         </Switch>
       </BrowserRouter>
