+++ src/Components/Login.js
@@ -1,4 +1,4 @@
-import React from "react"
+import React, { useState } from "react"
 import {
   Avatar,
   Button,
@@ -33,16 +33,51 @@ const useStyles = makeStyles((theme) => ({
   },
 }))
 
-export default function Login({
-  loginhandler,
-  inputName,
-  inputpassword,
-  nameChangeHandler,
-  passwordChangeHandler,
-  redirect,
-  alert,
-}) {
+export default function Login({ setIsAdmin }) {
   const classes = useStyles()
+  const [inputName, setInputName] = useState("")
+  const [inputpassword, setinputpassword] = useState("")
+  const [redirect, setRedirect] = useState(false)
+  const [alert, setAlert] = useState(false)
+
+  function userLogin() {
+    if (inputName === "asim" && inputpassword === "asim") {
+      return true
+    }
+    return false
+  }
+  function adminLogin() {
+    if (inputName === "admin" && inputpassword === "admin") {
+      return true
+    }
+    return false
+  }
+
+  function loginhandler(e) {
+    e.preventDefault()
+    if (userLogin()) {
+      //redirected to App.js
+      setIsAdmin(false)
+      setRedirect(true)
+      setAlert(false)
+    }
+    //  return <Redirect to="/todo" true/>;
+    else if (adminLogin()) {
+      setIsAdmin(true)
+      setRedirect(true)
+      setAlert(false)
+      //  return <Redirect to="/todo" true/>
+    } else {
+      setAlert(true)
+    }
+  }
+  function nameChangeHandler(e) {
+    setInputName(e.target.value)
+  }
+
+  function passwordChangeHandler(e) {
+    setinputpassword(e.target.value)
+  }
 
   return (
     <>
@@ -93,10 +128,12 @@ export default function Login({
               >
                 Sign In
               </Button>
-              {alert && (
+              {alert ? (
                 <Alert variant="outlined" severity="error">
                   Invalid Credentials! Please Try Again
                 </Alert>
+              ) : (
+                <></>
               )}
             </form>
           </Box>
