+++ src/theme/theme.js
@@ -1,0 +1,9 @@
+import { orange, red } from "@material-ui/core/colors"
+import { createMuiTheme } from "@material-ui/core"
+
+export const theme = createMuiTheme({
+  palette: {
+    primary: orange,
+    secondary: red,
+  },
+})
